Add files via upload
Insights
The bill amount ranges from around 3 dollars to 50 dollars while the average bill amount is around 20 dollars. Half of the bill amounts are less than 18 dollars.
The tip amount ranges from around 1 dollar to 10 dollars. The mean and median of the tip amounts are very close to each other (~3 dollars).
The group size varies from 1 person to 6 people.
50% of the people pay less than 20 dollars for overall bill and 50% of the people pay less than 3 dollars as tip.
There is a linear relationship between total_bill and tip i.e higher the billing amount, higher is the tip amount and vice versa.
The relationship between total_bill and tip is linear with respect to both the genders, day and time
Non smokers show a more prominent linear relationship between total_bill and tip as compared to smokers
With respect to size of the group, this relationship between total_bill and tip gets constant as the size of the group increases.
The number of orders is higher on weekends as compared to the weekdays. The order count is higher during the dinner time as compared to lunch
The median billing amount on Saturdays and Sundays is higher than the other two days.
The median tip amount is almost equal during Friday, Saturday and Sunday and the median during these three days is higher as compared to Thursday
50% of the customers spend more than 19 dollars on the overall bill during the dinner time as compared to 16 dollars during lunch
The median tip amount given during dinner is almost 1 dollar higher than that given during lunch time.
The number of male and female bill payers is around 160 and 80 respectively i.e the number of male bill payers is almost double as compared to the number of female bill payers
On Saturdays and Sundays, the number of male bill payers is significantly higher than the number of female bill payers.
The median value for tip is almost equal for both the genders.
Males have a lot of outliers i.e far away values, giving higher tips.
The tables which included non-smokers is around 60 more than those which included smokers
The number of non-smokers is significantly higher than the number of smokers on Thursdays and Sundays
Among all the four days, only on Fridays, the number of smokers is higher than the number of non-smokers
There is no relation between smoking status and tips, the median is almost the same in both the classes